backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Machine Learning & NLP
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LangChain & RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-aws>=0.2.0
langchain-openai>=0.1.0

# Vector Store
faiss-cpu>=1.7.4

# PDF Processing
pypdf>=4.0.0

# AWS Services
boto3>=1.28.0
botocore>=1.31.0

# OpenAI Services
openai>=1.0.0

# Google Gemini
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Additional Dependencies
python-multipart>=0.0.6
